fix: verify associate in product and others relations tables

+++ libs/category/src/useCase/deleteCategory/delete-category.service.ts
@@ -1,7 +1,12 @@
 /* eslint-disable no-param-reassign */
-import { UpdateResult } from 'typeorm';
-
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { IProductRepository } from '@product/product/implementations/product.interface';
+import { ProductRepository } from '@product/product/infra/typeORM/repositories/product.repository';
 
 import { ICategoryRepository } from '../../implementations/category.interface';
 import { CategoryRepository } from '../../infra/typeORM/repositories/category.repository';
@@ -11,6 +16,8 @@ export class DeleteCategoryService {
   constructor(
     @Inject(CategoryRepository)
     private categoryRepository: ICategoryRepository,
+    @Inject(ProductRepository)
+    private productRepository: IProductRepository,
   ) {}
 
   async execute(uuid: string): Promise<void> {
@@ -23,6 +30,15 @@ export class DeleteCategoryService {
     }
 
     const CategoryId = CategoryExists.id;
+
+    const FindConflictInCategoryAndProduct =
+      await this.productRepository.findByCategoryId(CategoryId);
+
+    if (FindConflictInCategoryAndProduct) {
+      throw new ConflictException(
+        'Não é possível excluir. Categoria associada a um ou mais produtos',
+      );
+    }
 
     return this.categoryRepository.delete(CategoryId);
   }
